requirements.txt
roboflow=0.2.34
streamlit=1.20.0
requests=2.28.2 
pandas=1.5.3
Pillow=9.4.0
python-dotenv=1.0.0

